WIP test case

+++ crates/rustc_plugin/tests/test_example.rs
@@ -1,10 +1,30 @@
-use std::{env, fs, path::Path, process::Command, sync::Once};
+use std::{
+  env, fs,
+  path::Path,
+  process::{Command, Output},
+  sync::Once,
+};
 
 use anyhow::{ensure, Context, Result};
 
 static SETUP: Once = Once::new();
 
 fn run(dir: &str, f: impl FnOnce(&mut Command)) -> Result<()> {
+  let output = run_configures(dir, true, f)?;
+
+  ensure!(
+    output.status.success(),
+    "Process exited with non-zero exit code"
+  );
+
+  Ok(())
+}
+
+fn run_configures(
+  dir: &str,
+  remove_target: bool,
+  f: impl FnOnce(&mut Command),
+) -> Result<Output> {
   let root = env::temp_dir().join("rustc_plugin");
 
   let heredir = Path::new(".").canonicalize()?;
@@ -42,12 +62,11 @@ fn run(dir: &str, f: impl FnOnce(&mut Command)) -> Result<()> {
 
   f(&mut cmd);
 
-  let _ = fs::remove_dir_all(ws.join("target"));
+  if remove_target {
+    let _ = fs::remove_dir_all(ws.join("target"));
+  }
 
-  let status = cmd.status().context("Process failed")?;
-  ensure!(status.success(), "Process exited with non-zero exit code");
-
-  Ok(())
+  cmd.output().context("Process failed")
 }
 
 #[test]
@@ -65,4 +84,14 @@ fn basic_with_arg() -> Result<()> {
 #[test]
 fn multi() -> Result<()> {
   run("workspaces/multi", |_cmd| {})
+}
+
+#[test]
+fn caching() -> Result<()> {
+  let workspace = "workspaces/basic";
+  let first_run = run_configures(workspace, false, |_| {})?;
+
+  let second_run = run_configures(workspace, true, |_| {})?;
+  ensure!(first_run == second_run);
+  Ok(())
 }
